consolidated Dpkg and parser setup into a collection resource
